initial itemAttributes functionality
added a reference number corresponding to # in corresponding sprite
sheet for every item, as well as a string indicating size. PlayerSprites
now refers to the reference number when changing sprites, be sure to
find the appropriate sprite in "items_lefthand" and enter it into the
InHandReference field in the itemAttributes section of its prefab. Also
items now have sizes which are edited similarly. Small objects are the
only ones that can fit in pockets- try it out.