Added PlayerCollisions and experimented with use
I experimented for a long while trying to figure out how to move
someone elses rigidbody over PUN. Turns out that the owners photonview
will keep putting the object back to where it believes it should be,
making it impossible to push it or move it. Same goes for transforms.
Need to research a solution