Implemented Access Denied anim and soundfx
Trying to open a door with at least 1 restriction without a card in the
ID slot or in the hand will trigger the Access Denied action. Only
works for doors that do not use the animator (need to handle doors with
animator in future).

Also fixed the inhand sprite problem for the general ID card (other
types will need to be fixed in the future)