Reworked the map editor.
Added some code on the map editor views to recursivelly search for tile
prefabs on 3 folders : Objects, Items and Structures.
Changed the file Structures so now adding a folders and tiles will
automatically load a representations on the map editor.
Code is not optimized, looking forward for tips on how make it less
repetitive.