Fixes: Closet inventory, UI equip roll out, Default item sprite order
Closet Inventory bug fixes:
- Fixed a problem where the inventory was referencing a matrix list
when it should of been receiving a copy
- Fixed a problem where items where not being removed from the last
tile they were on when being added to the closet. This made it possible
to steal items from cupboards by opening and closing a cupboard on the
last known position of the item

UI bug fixes:
- None of the equipment slots in the Equip load out were interactable.
This was because the slots were children of the main equipment button.
Moved the children up one layer

Default sprite order bug:
- If SpriteRenderer for an item was left at the default spriteOrder(0)
then the item would not be interactable when placed on a table before
runtime. Implemented a check to see if it is 0 on start and the change
it to 1 if it is. This should make sure the same mistake cannot happen
again in the future