Added RestrictiveTileMove property to matrix
Added:
- A new property to check whether the player is on a tile that is
restricting certain directions. This now makes it possible to create
glass sliding doors and allow players to crawl into windows that have
had the grill removed.
- Added CellTimer wallmount
- Added WinDoorAnimator placeholder

Changes:
- Removed the base class DoorController.cs from WinDoorController as
the behaviours of the WinDoor are completely different from a normal
door. The major being that you click to open a glass door, not run into
it
- Changed a the right windows in the cells to reinforced walls to place
the timers on
- Adjusted PlayerMove.cs to check for restrictive tiles