Major performance improvements and hiding player fixes
This commit doubles the FPS of the game build. It also improves the
player sync with server by removing the server updates from a SyncVar
and makes it use a sequenced realible channel instead. This also
completes the Hide in Cupboard feature for players.

Changes:
- Deactivated all redundant cameras (background and interact) and added
the layers back on main cam
- Added UpdateManager to control “tickrate” of objects with
Update/LateUpdate/FixedUpdate. For more info see the readme in
Script/Managers/UpdateManager/
- Move serverstate sync from a SyncVar Hook and to using ClientRpc’s
instead. This has improved the movement of other players on a client
but still there is a few hicups as the state doesn’t arrive in time
(player stops for a split second)
- The closetplayerhandler now moves the deactivated player with the
closet allowing the player to interact with the cupboard (jumping out
if it is unlocked while it is being pulled)