Crit Overlays part 1
Added:
- OverlayCritHole shader. It takes a texture and cuts out a hole in the
center. Fade off can also be adjusted to blur the edges
- Added a color property to the shader so it can be adjusted for
different scenarios (red for being attacked while unconscious, black
for almost dead etc)
- Added OverlayCrit component to handle the overlay texture. It will be
used to control the animations in part 2
- Added method to reset the overlay position when the UI changes
(resizes etc)
- Added Unconscious.psd. This is the texture that is being used as the
overlay. Made it a psd so it can be improved up on aesthetically (will
work on this more in part 2)

TODO (part 2):
- Create the conditions to trigger each animation. The texture will be
animated by controlling the properties of the shader itself
- Add the soundeffects (pumping heart, ambience)
- Tie it in with the health of the player