The road to responsiveness (UI Overhaul)
Changes:
- Removed the limitations to game client window size
- Changed the UI elements to rely more heavily on the Canvas Scaler.
Making it possible to have any window size and have the elements react
accordingly
- Changed the way equipment roll outs are hidden (changing the color
alpha channel to 0 and removing the raycast target. Doing it this way
side steps a refactor with the way itemSlots are handled when adding
items to them. They turn on and off the image component so changing
alpha was an easy away around that)

TODO:
- Some aspect ratios are really bad to the way sprites and images are
shown on screen, leaving lines between floor tiles and making UI
buttons super small or sometimes making them disappear all together.
Forcing a minimum aspect ratio will fix this.
- Adjust the camera size based on the size of the window. At the moment
you can see a big part of the map by hiding the right panel and going
full screen. I will make it so the camera can react to different sizes
so the same amount of screen is shown to everyone.