Finished implementation of overlay crit system
- Finished the UI_HealthMonitor component:
 The component will monitor the current state of the localplayer and
update the Right Hud heart monitor accordingly.
- Moved the crit soundfx to the UI_HeartMonitor
- Added a crude calculation for PlayerHealth maxHealth which can be
expanded later for other damage methods. Atm it just calculates the
blood level into a percentage and then applied to maxHealth.

All the systems are built to be easily customizable. The settings for
the hole overlay can be changed via the inspector and more settings can
be added easily. The healthmonitor anim is created in a coroutine and
the sprite starts can be adjusted via the inspector also.