Added PlayerHealthReporting and refactored parts of the health system
- Added PlayerHealthReporting component to handle the server —> client
communications for individual player objects. The idea is to have the
reporting component perform all of the UI updates via a call from the
server and to handle all player health queries from other players or
machines
- Started a big refactor of the health elements to simplify the system
and also the UI hierarchy to make it easier to scale up in the future.

TODO:
There seems to be network updates happening within healthbehaviour that
will need to be removed and allow the HealthReporting component to
handle all of the network updates (I.E atm if you kill someone fast
enough the UI crits does not have time to update because it relies on
another system that seems to be a duplicate of the underlying
HealthBehaviour)