I love you, Meat Steak.
Adds an interaction for edible food.  BaseFood.cs script is added and small changes to objectManager.cs to include food sprites.  MeatSteak.cs is included with prefab in Outpost Deathmatch that shows how to change players health effects and how to play a sound on the SoundNetworkManager where the player is currently standing when they devour that tasty slutty meat steak. More info in BaseFood.cs class summary.