Added a specialty clause to the licence
This addition to the licence, gives us the possibility to use closed-source libraries, without any future fear of licence-virality. This for example, makes it possible to have more integration with libraries like the steam SDK.

Although some regions, like europe, prevent code virality by law, not all do so. In common law systems like the US, it's not very clear if the FAQ from GNU/FSF takes precedence over the literal text of the AGPL licence. To prevent any future (and/or current) issues with licensing, this rule is put in place.