Fixes #1373 - fix blood splatter effect so it appears underneath the mob rather than off to the side.
The BloodSplatter method was trying to instantiate the BloodSplatter GameObject under a parent. The parent caused the blood splatter to be offset.
There doesn't seem to be a need to have the splatter created underneath the parent (none of the other effects do it when they call PoolNetworkInstantiate) so I think it's fine to leave the parent as null.