Added DestroyItemInSlot method on InventoryManager
To make it more clear about which method can Destroy items in a slot I've made a new public method on the inventory manager with documentation named DestroyItemInSlot(). Use this method for items that are meant to be consumed while in an inventory slot. Server only command.

Also renamed similiar method DisposeItemServer to DestroyItemInSlot that can take an item instead of an inventory slot parameter and updated references.