Fix for issue #1614
ControlDisplays now ensures that the job select window is inactive at the start of the round. Also added fallback to ensure that only one of either the job select window or team select window can be visible at one time and only if the player doesn't already have a job.