Changes walking check in Slip(), and adds summaries to Slip() and Stun().
Slip() now checks if the player is walking by checking if the player's speed is greater than walking speed. Before this Slip() used the UIManager to check if the player is walking.
Added summaries and parameter descriptions to Slip() and Stun().
Added an option to slip while walking in Slip().
Players now always slip on bananas.