Adds rejoin system
Removes HandleDisconnect() and it's call in CustomNetworkManager.cs.
Moves part of RespawnPlayer() to a new function: TransferPlayer().
Adds List<ConnectedPlayer> loggedOff in PlayerList.cs. When players are removed in TryRemove() they are added to the LoggedOff list to be retrieved when they relog.
Added TakeLoggedOffPlayer() in PlayerList.cs which returns a logged player with a certain steamId and removes them from the list.
Refactored JoinedViewer's player setup so the client sends a steamId to the server, the server does serverside player setup with the steamId, then tells the client to do clientside player setup with the character they logged out with (if they logged out).
Added CmdRejoin() in JoinedViewer which lets the player rejoin their character.
Adds ReenterBodyUpdates(), SendSyncMessage(), and UpdateInventorySlots() to PlayerNetworkActions.cs, which are used to sync a rejoined player.
Creates setLocalPlayer() in PlayerSync.cs from existing setLocalPlayer code so it can be used with the rejoin system.
Stops the player gameobject from being destroyed on logout.
Gives CmdServerSetupPlayer() "0" (ConnectedPlayer.steamId's default value) as an argument if it is not a steam server.