Fixes character settings not passing on if the player is logged out while transfering bodies.
Removes the automatic relog toggle copypaste, now the GUI of the login screen will just check for the preferences and run the standard login.
Character settings will now only transfer from the client to the server when the player is a JoinedViewer, on other type of body transfers, the settings will move from body to body.
The CharacterSettings is now saved in PlayerScript.