Adds empty hand interactions and makes various changes/fixes
Fixes knifes and scalpels not working on animals/players.
Adds CPR.
Adds disarms.
Adds hugs.
Adds punches.
Adds the method InventoryManager.DropItemInSlot(). This drops the item in a specific slot (needed for disarm).
Changes PostToChatMessage.SendItemAttackMessage to SendAttackMessage. This makes it more generic so it can now be used for punches and other attacks in the future.