General work on handcuffs
Cuffs now appear in left and right hand of the player when the player is handcuffed.

The existing code refered to a inventory slot named "Handcuffs" which has been added.

Sprite for left and right cuffs has been added as a property inside a monobehaviour object to the left and right hands UI slots.

Code has been added to set the secondary sprite of Left/Right hand UI slots by  getting the sprite from that property.

What remains to be done : rework the AlertUI to make it generic and accept a new alert UI action that will trigger the uncuff of the player.  I'm not proficient enough (yet) with the project and Unity to do that.  Please put back that feature in the To-Do list for a more pro to finish it (sorry 'bout that, I'm now more conscient of my Unity proficiency level)