Backpacks fitting in backpacks fixed
Base backpack prefab was set to tiny in it's item traits. Just set it to huge like the actual use backpacks. I think whoever posted the issue might have spawned the base prefabs and not the established ones. I recommend a different naming convention so that they can be differentiated in game's dev menu.