Additional Security Clothing and Modified Populators.
Adds more clothing for security, including the iconic armor vest and helmets, and changes populators for Security Officers so that they spawn with armor vests and helmets, and changes the populator for the security backpack so that they will have security gas masks in them. Also adjusts the properties of the the HoS's coat and beret.