Lets AirLockAnimator work with airlocks with no glass/light overlay.
AirLockAnimator was coupled to AirLocks with overlay sprites. This airlock script only works on airlock gameobjects with four children gameobjects containing spriterenders for the door, light overlay, glass/cover overlay, and weld overlay.

When the glass/cover overlay gameobject spriterender is empty, the AirLockAnimator takes every sprite in "Assets/Resources/" and adds it to the AirLockAnimator overlay sprites.
This commit adds a check to prevent this, and adds checks that prevent the overlay animations from playing when there are no sprites.