Third pass on bugs + Minor changes to PlayerLightControl
PlayerLightControl renamed to ItemLightControl, separated the Toggle method in ItemLightControl into Toggle and Intensity. Fixed the bugs that were caused by separating the methods, made some Minor changes to some of the PDA menus for text updating, failed to make notes sync, put a temporary blocker on it till we get someone that is better at networking to do it.