NRE Correction Issue #4690
NRE Correction Issue #4690

Correction of a server side NRE when a grenade explodes in a locker.
The grenade was despawned but still in the locker list of item.

Suggestion for a more intelligent software : the locker should not handle it's own list of item.  We should one day create a component for lockers, boxes, crates, and other containers.