Fixed atmos breaking on shooting guns
The problem was in the hotspot call which was called from registerTile of the gun. So if the shooter is on the shuttle matrix, gun was on a station matrix. Station matrix checks if there is atmos on teh tile from which gun shot, and there is of course none, so it was creating a new MetaDataNode with empty space atmos. So caching and calling hotspot from register tile component of the shooter fixes it.