Added a loading screen.
The loading screen prefab uses a new loading bar prefab, which is a modified slider component.
The screen also includes a TMP text that can be used to inform the player of what is currently being loaded.
It is placed at the last point in the UI manager, which means it renders on top of all other UI elements.
Tested against the standard resolutions and seems to scale alright.