Escaped prefix labels, changed scene labels
Requires labels to be updated before merging!

Escaped all keys in the label yaml. `js-yaml` *should* be able to handle this, but I only tested it in an online parser.

The escaped keys contain some changes. Map names now have a `"Scene:"` prefix to make sure the labels appear next to each other on the list.