Adjusts SeedExtractor Script
Changes SeedExtractor Script to dumb it's seed packets on despawn. However, for whatever reason this only works when the device is not powered, it will drop the produce you put in. If the SeedExtractor is powered, it will process the produce and this somehow removes the seedpackets from the seedpacket list even though it shouldn't, so when you deconstruct no packets come out.