Various Radial QoL features (#5920)
* Various Radial QoL features

Radial menu opens up the same as before. However, the menu itself no longer moves around on the screen as the player moves. Instead the branch will shrink/grow.

The branch origin now centers on the tile that is clicked.

If a tile instead of a UI element is clicked, the branch will show under the UI.

The branch can follow tiles on moving matrices. Matrix offset is not currently accounted for yet so there will be a jump as the matrix changes positions.

* Appease codacy