Adds UI sprite scaling, switch to leantween for radial menu highlighting
Many sprites can be hard to see/read in the UI and radial. This adds in processing for sprites to gather scaling and offset data without the need to modify all textures and sprites. UI images can apply sprite scaling with an extension method by passing it the sprite. Sprite/texture metadata is created lazily and cached.

Adds icon scaling for the radial menu.

Switches from using CrossFadeAlpha for highlighting in the radial menu and uses leantween instead.