Adds UI sprite scaling (#5968)
* Adds UI sprite scaling, switch to leantween for radial menu highlighting

Many sprites can be hard to see/read in the UI and radial. This adds in processing for sprites to gather scaling and offset data without the need to modify all textures and sprites. UI images can apply sprite scaling with an extension method by passing it the sprite. Sprite/texture metadata is created lazily and cached.

Adds icon scaling for the radial menu.

Switches from using CrossFadeAlpha for highlighting in the radial menu and uses leantween instead.

* Texture data null check.

* Forgot to update a name

* Adds assembly definition. Use pointers instead of copying.

Switches from copying to a pixel cache to using pointers to read textures instead.

Sneakily fix an issue with palettes that I must have broke at some point.

* in to ref