Change MatrixInfo from struct to class
MatrixInfo is a struct. A struct with 20+ fields. A struct that is copied in a lot of functions. This really should not be a struct.
Changing this to a class prevents copying all the fields. This is a significant time savings (around 30% for small functions) for anything that uses MatrixInfo.