Audio Parameters Were Not Serializable (#6033)
* Audio Parameters Were Not Serializable

Because of the use of nullable variables and get/setters, AudioSourceParameters and ShakeParameters could not be serialized, making PlaySoundMessage not work as intended.

* Aranclanos's fixes

* Changed Audio Parameters from Classes to Structs

This is a large change that significantly alters how AudioSourceParameters, ShakeParameters, and SoundManager's ApplyAudioSourceParameters works.  An additional method ForceAudioSourceParameters was added for better control.

Also, MobExplore was calling SoundManager incorrectly.

* Muffled Was The Default Audio Mixer

This removed undefined from the audio mixer settings and volumerollofftype, as they were unused, and changed the default mixer from muffled to master and the default rolloff to linear from logarithmic.