Change Reaction.cs to not rely on the serilization of nullable floats. Change ReactionEditor.cs to be compatible.  (#7093)
* Change Reaction.cs to not rely on serializing nullable floats and ReactionEditor.cs to be compatible with these changes. Use the new system on Ash.asset.

Serializing nullable floats is unsupported by unity and does not work. This commit makes it so that they are no longer used when modifying chemical reactions. This is achieved by adding two pairs of a float that stores the temperature value and a bool that stores if that value is null. Then, when apply is called, these values are moved to nullable floats.

* Hide the values in inspector so it looks nicer.

* Update ReactionTests.cs

Make the unit tests use the new system. (facepalm)