Changed menu UI as rq'd by ticket 7765
LOGOUT has been renamed to "Exit to Main Menu" with the font sized down
a bit to make space for the full text. The menu has been resized a
little bit to make the space above and below the new text consistent
with the other menu items. "Exit to main menu" has been moved down to
above Exit Game as requested by the ticket.

~ Cosmosgirlandcrows