Changes in the ingame menu UI as requested by ticket #7765 (#7924)
* Changed menu UI as rq'd by ticket 7765

LOGOUT has been renamed to "Exit to Main Menu" with the font sized down
a bit to make space for the full text. The menu has been resized a
little bit to make the space above and below the new text consistent
with the other menu items. "Exit to main menu" has been moved down to
above Exit Game as requested by the ticket.

~ Cosmosgirlandcrows

* LOGOUT text changed as requested by #7765

As requested by ticket #7765, LOGOUT in the ingame menu has been
replaced with "Exit to Main Menu". In order to uphold consistency in the
menu I have also changed "Exit Game" to "Exit to Desktop" and slightly
adjusted font size and button size to keep the style consistent.

* Changed logout confirm text