Moves VerifyChildReference into extension methods
Adds some ease of use features.

If no reference is found, it will log the component it is missing from, the object the component is located on, the reference name, a brief description, and the component type needed.

It now checks for a child with a given name, a child with a name that matches the reference method/property name, or a component that matches the type if there is only one.