Ensure verification only happens once.
If a reference is missing, a simple component is added to the offending object. This is to make sure that, in the event that a child component isn't found, accessing the component reference doesn't keep going through the hierarchy to find a component that doesn't exist. If a component has failed verification, it will only log the event and try to find a valid child component once. The component is only ever added if the reference fails verification.