Consolidate radials
Adds in the ability to use the same right click radial for a quick single ring custom radial with some tweaks.

Airlock painter uses the right click radial now (as a single ring without the branch) and removes the duplicated radial.

If the radial is opened and there is only one object on the tile, the actions for that object are placed directly on the main ring (no action ring). This includes action bar items. There is no delay when hovering over items when it is a single ring.

Fixes Airlock Science Assembly (it used the command door prefab).

Moves some right click related scriptable objects around and renames a couple things.