PogstationREDUX Patch v2 (#9721)
* PogstationREDUX Patch v2

Various patches for minor issues to PogstationREDUX.

* Adds fuel rods.

Adds fuel rods into the radioactive crate in nuclear chamber.

* General fixes to PSR

Adds all machines that could to the ore silo and likewise with the research server. (also places gibber in kitchen and places monkey cubes instead.

* Fixes the gibber in freezer.

Simply deletes the gibber and replaces it with a new one to avoid the APC errors.